pyproject.toml
[tool.poetry]
name = "repl_nix_ChatGPT-Line-Bot"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.3"
line-bot-sdk = "^2.4.1"
openai = "^0.27.0"
python-dotenv = "^1.0.0"
bs4 = "^0.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
